Добавлены новые функции для предзагрузки данных на странице Dashboard, включая сводную статистику и временные ряды. Обновлен компонент SummaryWidget для отображения иконок и улучшения стилей. Исправлены стили в TimeseriesWidget и CalendarPopover для лучшего отображения. Изменен импорт в DashboardPage для использования SummaryWidget.

+++ src/pages/dashboard/ui/ResentCompanies.tsx
@@ -74,7 +74,7 @@ export const ResentCompanies = () => {
 		return (
 			<Layout direction='column' className='gap-4'>
 				{data.items.map((campaign: CampaignListItem) => (
-					<CampaignCard key={campaign.id} campaign={campaign} variant='short' />
+					<CampaignCard key={campaign.id} campaign={campaign} variant='long' />
 				))}
 			</Layout>
 		)
